+++ level.c
@@ -14,11 +14,10 @@
  */
 
 #include <stdlib.h>
-#include <stdint.h>
-#include <string.h>
 #include <stdbool.h>
 
 #include <assert.h>
+#include <string.h>
 
 #include "err.h"
 #include "level.h"
@@ -26,17 +25,17 @@
 
 static struct level *_allocate(struct dimension _d);
 
-static void _generate_dungeon(struct level *_l, uint8_t _rooms,
+static void _generate_dungeon(struct level *_l, unsigned int _rooms,
     struct dimension _room_min, struct dimension _room_max);
 
 static void _carve_room(struct level *_level, struct coordinate *_anchor,
     struct dimension _room_min, struct dimension _room_max);
 
 static void _connect_anchors(
-    struct level *_level, uint8_t _rooms, struct coordinate _anchors[]);
+    struct level *_level, unsigned int _rooms, struct coordinate _anchors[]);
 
 struct level *
-level_create(struct dimension d, uint8_t rooms, struct dimension room_min,
+level_create(struct dimension d, unsigned int rooms, struct dimension room_min,
     struct dimension room_max)
 {
 	assert(d.height > 0);
@@ -58,7 +57,7 @@ level_destroy(struct level *level)
 {
 	assert(level != NULL);
 
-	for (uint8_t y = 0; y < level->dimension.height; y++) {
+	for (unsigned int y = 0; y < level->dimension.height; y++) {
 		free(level->tiles[y]);
 	}
 
@@ -68,12 +67,12 @@ level_destroy(struct level *level)
 
 void
 level_modify_random_floor_tiles(
-    struct level *level, uint8_t count, uint8_t mask)
+    struct level *level, unsigned int count, unsigned int mask)
 {
-	for (uint8_t i = 1; i < count; i++) {
+	for (unsigned int i = 1; i < count; i++) {
 		for (;;) {
-			uint8_t y = rand() % (level->dimension.height);
-			uint8_t x = rand() % (level->dimension.width);
+			unsigned int y = rand() % (level->dimension.height);
+			unsigned int x = rand() % (level->dimension.width);
 
 			if (level->tiles[y][x] & TA_FLOOR) {
 				level->tiles[y][x] |= mask;
@@ -100,7 +99,7 @@ _allocate(struct dimension d)
 
 	assert(l->tiles != NULL);
 
-	for (uint8_t y = 0; y < l->dimension.height; y++) {
+	for (unsigned int y = 0; y < l->dimension.height; y++) {
 		l->tiles[y] = calloc(l->dimension.width, sizeof(**l->tiles));
 		if (l->tiles[y] == NULL)
 			err("level _allocate: calloc");
@@ -123,14 +122,14 @@ _allocate(struct dimension d)
  *    the anchor of the next room.
  */
 static void
-_generate_dungeon(struct level *level, uint8_t rooms, struct dimension room_min,
-    struct dimension room_max)
+_generate_dungeon(struct level *level, unsigned int rooms,
+    struct dimension room_min, struct dimension room_max)
 {
 	struct coordinate anchors[rooms];
-	for (uint8_t i = 0; i < rooms; i++)
+	for (unsigned int i = 0; i < rooms; i++)
 		_carve_room(level, &anchors[i], room_min, room_max);
 
-	for (uint8_t i = 0; i < rooms - 1; i++)
+	for (unsigned int i = 0; i < rooms - 1; i++)
 		_connect_anchors(level, rooms, anchors);
 }
 
@@ -138,22 +137,23 @@ static void
 _carve_room(struct level *level, struct coordinate *anchor,
     struct dimension room_min, struct dimension room_max)
 {
-	uint8_t height = (rand() % (room_max.height - room_min.height + 1)) +
+	unsigned int height =
+	    (rand() % (room_max.height - room_min.height + 1)) +
 	    room_min.height;
 
-	uint8_t width =
+	unsigned int width =
 	    (rand() % (room_max.width - room_min.width + 1)) + room_min.width;
 
-	uint8_t oy = rand() %
+	unsigned int oy = rand() %
 	    (level->dimension.height - height + 1 /* full height */ -
 		2 /* borders */);
 
-	uint8_t ox = rand() %
+	unsigned int ox = rand() %
 	    (level->dimension.width - width + 1 /* full width */ -
 		2 /* borders */);
 
-	for (uint8_t y = 0; y < height; y++) {
-		for (uint8_t x = 0; x < width; x++)
+	for (unsigned int y = 0; y < height; y++) {
+		for (unsigned int x = 0; x < width; x++)
 			level->tiles[oy + y + 1][ox + x + 1] = TA_FLOOR;
 	}
 
@@ -163,33 +163,33 @@ _carve_room(struct level *level, struct coordinate *anchor,
 
 static void
 _connect_anchors(
-    struct level *level, uint8_t rooms, struct coordinate anchors[])
+    struct level *level, unsigned int rooms, struct coordinate anchors[])
 {
-	for (uint8_t i = 0; i < rooms - 1; i++) {
+	for (unsigned int i = 0; i < rooms - 1; i++) {
 		struct coordinate start = anchors[i];
 		struct coordinate stop = anchors[i + 1];
 
-		uint8_t ay = start.y;
-		uint8_t by = stop.y;
+		unsigned int ay = start.y;
+		unsigned int by = stop.y;
 
 		if (ay > by) {
 			ay = stop.y;
 			by = start.y;
 		}
 
-		uint8_t ax = start.x;
-		uint8_t bx = stop.x;
+		unsigned int ax = start.x;
+		unsigned int bx = stop.x;
 
 		if (ax > bx) {
 			ax = stop.x;
 			bx = start.x;
 		}
 
-		for (uint8_t x = ax; x <= bx; x++) {
+		for (unsigned int x = ax; x <= bx; x++) {
 			level->tiles[start.y][x] = TA_FLOOR;
 		}
 
-		for (uint8_t y = ay; y <= by; y++) {
+		for (unsigned int y = ay; y <= by; y++) {
 			level->tiles[y][stop.x] = TA_FLOOR;
 		}
 	}
